Fixed bug #167
The edit command will now also iterate through all the assigned tasks of
the edited employee. For each of this tasks, iterate through all the
tasks in the model and look for the task itself.

After this then access the assigned employees of each of those tasks and
remove and re add this new edited employee.

This has to be done because the assigned employees and assigned tasks
contains an immutable object of task or employee that is not the same as the one stored in
the main model object for the app. This is why the error happens to
begin with.